feat: init client-sync

+++ src/voyage/services/voyage.service.ts
@@ -1,14 +1,23 @@
-import { Inject, Injectable } from '@nestjs/common';
+import {
+  BadRequestException,
+  Inject,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
+import { EventBus } from '@nestjs/cqrs';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository, In, DataSource } from 'typeorm';
 import { POSTGRES_DATA_SOURCE } from 'src/database/constants';
-import { User } from '../auth/user.model';
-import { Passenger } from '../passenger/models/passenger.model';
-import { Ship } from '../ship/models/ship.model';
-import { CreateVoyageDto } from './dto/create-voyage.dto';
-import { UpdateVoyageDto } from './dto/update-voyage.dto';
-import { VoyageCrew } from './models/voyage-crew.model';
-import { Voyage } from './models/voyage.model';
+import { User } from '../../auth/models/user.model';
+import { Passenger } from '../../passenger/models/passenger.model';
+import { Ship } from '../../ship/models/ship.model';
+import { CreateVoyageDto } from '../dto/create-voyage.dto';
+import { UpdateVoyageDto } from '../dto/update-voyage.dto';
+import { VoyageOnboardingCreatedEvent } from '../events/voyage-onboarding-created.event';
+import { VoyageCrew } from '../models/voyage-crew.model';
+import { Voyage, VoyageStatus } from '../models/voyage.model';
+
+// TODO: Fix call to other module
 
 @Injectable()
 export class VoyageService {
@@ -30,6 +39,8 @@ export class VoyageService {
 
     @InjectRepository(User)
     private readonly userRepository: Repository<User>,
+
+    private readonly eventBus: EventBus,
   ) {}
 
   async create(createVoyageDto: CreateVoyageDto) {
@@ -107,9 +118,15 @@ export class VoyageService {
       where: { id },
       relations: ['ship', 'guests', 'crews'],
     });
+
     if (!voyage) {
-      return null;
+      throw new NotFoundException('Voyage not found');
     }
+
+    if (voyage.status != VoyageStatus.DRAFT) {
+      throw new BadRequestException('Only draft voyages can be updated');
+    }
+
     if (updateVoyageDto.shipId) {
       const ship = await this.shipRepository.findOne({
         where: { id: updateVoyageDto.shipId },
@@ -133,6 +150,34 @@ export class VoyageService {
   }
 
   async remove(id: number) {
+    const voyage = await this.voyageRepository.findOne({ where: { id } });
+
+    if (!voyage) {
+      throw new NotFoundException('Voyage not found');
+    }
+
+    if (voyage.status != VoyageStatus.DRAFT) {
+      throw new BadRequestException('Only draft voyages can be deleted');
+    }
+
     return this.voyageRepository.delete(id);
+  }
+
+  async startOnboarding(id: number) {
+    const voyage = await this.voyageRepository.findOne({ where: { id } });
+
+    if (!voyage) {
+      throw new NotFoundException('Voyage not found');
+    }
+
+    if (voyage.status != VoyageStatus.SCHEDULED) {
+      throw new BadRequestException('Only scheduled voyages can be onboarded');
+    }
+
+    voyage.status = VoyageStatus.ONGOING;
+
+    this.eventBus.publish(new VoyageOnboardingCreatedEvent(voyage.id));
+
+    return this.voyageRepository.save(voyage);
   }
 }
